+++ src/libsail-common/src/resolution.c
@@ -35,7 +35,7 @@ sail_status_t sail_alloc_resolution(struct sail_resolution **resolution) {
     return SAIL_OK;
 }
 
-sail_status_t sail_alloc_resolution_from_data(struct sail_resolution **resolution, enum SailResolutionUnit unit, uint16_t x, uint16_t y) {
+sail_status_t sail_alloc_resolution_from_data(struct sail_resolution **resolution, enum SailResolutionUnit unit, float x, float y) {
 
     SAIL_CHECK_RESOLUTION_PTR(resolution);
